refactor(#457): split lifestyle.ts (236 lines) into lifestyleData/ submodules
